Added support for further shortening object paths in Ninja generator to work around the MAX_PATH limitation. Previously, the Ninja generator would try to shorten long object file paths as follows:
<build dir>/CMakeFiles/<target>/relative/path/source.c.obj => <build dir>/CMakeFiles/<target>/<MD5>/source.c.obj

If the build directory and target name were too long for this to work, the generator would proceed with the original path.
This patch allows shortening it further to the following form:

<build dir>/CMakeFiles/<target>/<MD5>

It is only used as a fallback when the regular shortening didn't work and is helpful for projects with relatively long file names.